Code with some erros

+++ drugdrop/src/main/java/com/uatqs/drugdrop/model/Drug.java
@@ -33,7 +33,7 @@ public class Drug {
         return this.id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
